+++ src/components/IngredientSearchInput/IngredientSearchInput.js
