+++ src/components/NutritionalDataSection/NutritionalDataSection.js
